requirements.txt
Django>=1.5.5
south
psycopg2==2.4.5
django-admin-bootstrapped
django-floppyforms
django-crispy-forms
django-guardian
django-json-field
django-leaflet
django-geojson
django-braces
django-simple-history
python-dateutil
django-tastypie

# Dev
django-debug-toolbar
django-rosetta
lxml # Tests - needs libxml2-dev libxslt-dev packages
